Make the failing negative case test pass
In making the negative case pass, it was necessary to add extra
code to the positive test for the for the coin collector.

In addition the interface of the coin collector was modified to return
a value when the coin was successfully added.

Finally changed the VS Code Unit Test runner to a more reliable
extension.